@@ -62,9 +62,9 @@
   },
   "dependencies": {
     "axios": "^0.18.0",
+    "body-parser": "^1.18.3",
     "bplist-parser": "^0.1.1",
-    "electron-oauth2": "^3.0.0",
-    "electron-osx-prompt": "^1.3.10",
+    "express": "^4.16.3",
     "motion-core": "https://github.com/motioncrypto/motion-core",
     "vue": "^2.5.13",
     "vue-electron": "^1.0.6",
@@ -93,8 +93,8 @@
     "css-loader": "^0.28.4",
     "del": "^3.0.0",
     "devtron": "^1.4.0",
-    "electron": "^1.7.5",
-    "electron-builder": "^19.19.1",
+    "electron": "^2.0.5",
+    "electron-builder": "^20.26.0",
     "electron-debug": "^1.4.0",
     "electron-devtools-installer": "^2.2.0",
     "eslint": "^4.4.1",
